Decorator雛形
