+++ src/Bold.java
@@ -8,51 +8,11 @@ import java.awt.event.ActionListener;
 /**
  * @author zx328
  */
-public class Bold implements ActionListener {
-    JTextPane textPane;
-    static JButton boldButton;
-    DefineImageButton defineImageButton;
-    UI ui;
+public class Bold extends TextFontDecorator {
+
     public Bold(JTextPane jtextPane){
         textPane=jtextPane;
-        defineImageButton = new DefineImageButton();
-        createBoldButton();
-    }
-    public void createBoldButton(){
-        //粗體
-        boldButton = new JButton(defineImageButton.boldIcon);
-        boldButton.setToolTipText("粗體");
-        boldButton.setText("bold");
-        boldButton.setFont(new Font("bold",0,0));
-        boldButton.addActionListener(this);
-    }
-    public static final void setCharacterAttributes(JEditorPane editor,
-                                                    AttributeSet attr, boolean replace) {
-        int p0 = editor.getSelectionStart();
-        int p1 = editor.getSelectionEnd();
-        if (p0 != p1) {
-            StyledDocument doc = getStyledDocument(editor);
-            doc.setCharacterAttributes(p0, p1 - p0, attr, replace);
-        }
-        StyledEditorKit k = getStyledEditorKit(editor);
-        MutableAttributeSet inputAttributes = k.getInputAttributes();
-        if (replace) {
-            inputAttributes.removeAttributes(inputAttributes);
-        }
-        inputAttributes.addAttributes(attr);
-    }protected static final StyledEditorKit getStyledEditorKit(JEditorPane e) {
-        EditorKit k = e.getEditorKit();
-        if (k instanceof StyledEditorKit) {
-            return (StyledEditorKit) k;
-        }
-        throw new IllegalArgumentException("EditorKit must be StyledEditorKit");
-    }
-    protected static final StyledDocument getStyledDocument(JEditorPane e) {
-        Document d = e.getDocument();
-        if (d instanceof StyledDocument) {
-            return (StyledDocument) d;
-        }
-        throw new IllegalArgumentException("document must be StyledDocument");
+
     }
 
     @Override
@@ -67,8 +27,6 @@ public class Bold implements ActionListener {
             StyleConstants.setBold(sas, bold);
             setCharacterAttributes(textPane, sas, false);
         }
-
-
 
     }
 }
