+++ src/Bottomline.java
@@ -7,51 +7,14 @@ import java.awt.event.ActionListener;
 /**
  * @author zx328
  */
-public class Bottomline implements ActionListener {
-    JTextPane textPane;
-    static JButton bottomlineButton;
-    DefineImageButton defineImageButton;
+public class Bottomline extends TextFontDecorator{
+
     public Bottomline(JTextPane jtextPane){
         textPane=jtextPane;
-        defineImageButton =new DefineImageButton();
-        createBottomLine();
-    }
-    public void createBottomLine(){
-        bottomlineButton = new JButton(defineImageButton.bottomlineIcon);
-        bottomlineButton.setToolTipText("底線");
-        bottomlineButton.setText("bottomline");
-        bottomlineButton.setFont(new Font("bottomline", 0, 0));
-        bottomlineButton.addActionListener(this);
+
 
     }
-    public static final void setCharacterAttributes(JEditorPane editor,
-                                                    AttributeSet attr, boolean replace) {
-        int p0 = editor.getSelectionStart();
-        int p1 = editor.getSelectionEnd();
-        if (p0 != p1) {
-            StyledDocument doc = getStyledDocument(editor);
-            doc.setCharacterAttributes(p0, p1 - p0, attr, replace);
-        }
-        StyledEditorKit k = getStyledEditorKit(editor);
-        MutableAttributeSet inputAttributes = k.getInputAttributes();
-        if (replace) {
-            inputAttributes.removeAttributes(inputAttributes);
-        }
-        inputAttributes.addAttributes(attr);
-    }protected static final StyledEditorKit getStyledEditorKit(JEditorPane e) {
-        EditorKit k = e.getEditorKit();
-        if (k instanceof StyledEditorKit) {
-            return (StyledEditorKit) k;
-        }
-        throw new IllegalArgumentException("EditorKit must be StyledEditorKit");
-    }
-    protected static final StyledDocument getStyledDocument(JEditorPane e) {
-        Document d = e.getDocument();
-        if (d instanceof StyledDocument) {
-            return (StyledDocument) d;
-        }
-        throw new IllegalArgumentException("document must be StyledDocument");
-    }
+
     @Override
     public void actionPerformed(ActionEvent e) {
         String s=e.getActionCommand();
@@ -64,15 +27,5 @@ public class Bottomline implements ActionListener {
             StyleConstants.setUnderline(sas, underline);
             setCharacterAttributes(textPane, sas, false);
         }
-        if(s.equals("bold")) {
-            StyledEditorKit kit = getStyledEditorKit(textPane);
-            MutableAttributeSet attr = kit.getInputAttributes();
-            boolean bold = (StyleConstants.isBold(attr)) ? false : true;
-            SimpleAttributeSet sas = new SimpleAttributeSet();
-            StyleConstants.setBold(sas, bold);
-            setCharacterAttributes(textPane, sas, false);
-        }
-
-
     }
 }
