+++ src/Italics.java
@@ -7,50 +7,13 @@ import java.awt.event.ActionListener;
 /**
  * @author zx328
  */
-public class Italics implements ActionListener {
-    JTextPane textPane;
-    static JButton italicsButton;
-    DefineImageButton defineImageButton;
+public class Italics extends TextFontDecorator  {
+
     public Italics(JTextPane jtextPane){
         textPane=jtextPane;
-        defineImageButton =new DefineImageButton();
-        createItalicsButton();
+
     }
-    public void createItalicsButton(){
-        italicsButton = new JButton(defineImageButton.italicsIcon);
-        italicsButton.setToolTipText("斜體");
-        italicsButton.setText("italics");
-        italicsButton.setFont(new Font("italics", 0, 0));
-        italicsButton.addActionListener(this);
-    }
-    public static final void setCharacterAttributes(JEditorPane editor,
-                                                    AttributeSet attr, boolean replace) {
-        int p0 = editor.getSelectionStart();
-        int p1 = editor.getSelectionEnd();
-        if (p0 != p1) {
-            StyledDocument doc = getStyledDocument(editor);
-            doc.setCharacterAttributes(p0, p1 - p0, attr, replace);
-        }
-        StyledEditorKit k = getStyledEditorKit(editor);
-        MutableAttributeSet inputAttributes = k.getInputAttributes();
-        if (replace) {
-            inputAttributes.removeAttributes(inputAttributes);
-        }
-        inputAttributes.addAttributes(attr);
-    }protected static final StyledEditorKit getStyledEditorKit(JEditorPane e) {
-        EditorKit k = e.getEditorKit();
-        if (k instanceof StyledEditorKit) {
-            return (StyledEditorKit) k;
-        }
-        throw new IllegalArgumentException("EditorKit must be StyledEditorKit");
-    }
-    protected static final StyledDocument getStyledDocument(JEditorPane e) {
-        Document d = e.getDocument();
-        if (d instanceof StyledDocument) {
-            return (StyledDocument) d;
-        }
-        throw new IllegalArgumentException("document must be StyledDocument");
-    }
+
     @Override
     public void actionPerformed(ActionEvent e) {
         String s=e.getActionCommand();
@@ -63,8 +26,6 @@ public class Italics implements ActionListener {
             StyleConstants.setItalic(sas, italic);
             setCharacterAttributes(textPane, sas, false);
         }
-
-
 
     }
 }
